requirements.txt
# Core Streamlit UI dependencies
streamlit>=1.28.0

# Existing agent dependencies (from your current setup)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40
pydantic>=2.0.0
typing-extensions>=4.0.0

# Web search and research tools
tavily-python>=0.3.0

# Display and formatting
rich>=13.0.0

# HTTP requests
requests>=2.31.0
